#39: Improved test coverage, moving interfaces to types
